[refactor] 검색바 컴포넌트 분리
