+++ src/components/Search/SearchBar.module.css
@@ -1,0 +1,25 @@
+.inputContainer {
+  display: flex;
+  flex-direction: row;
+  align-items: center;
+  margin-right: 24px;
+  width: 300px;
+  height: 36px;
+  border: 1px solid #dcdcdc;
+  border-radius: 5px;
+  /* transition: border-color 0.3s ease-in-out; */
+}
+
+.iconSearch {
+  margin-left: 12px;
+  color: #a4a4a4;
+}
+
+.searchBar {
+  width: 250px;
+  border: none;
+  outline: none;
+  border-radius: 6px;
+  padding: 0px 6px;
+  font-size: 14px;
+}
