+++ src/components/Search/SearchBar.tsx
@@ -1,0 +1,33 @@
+// SearchBar.tsx
+import React, { useState } from "react";
+import styles from "@src/components/Search/SearchBar.module.css";
+import { BiSearch } from "react-icons/bi";
+
+interface SearchBarProps {
+  onSearch: (searchTerm: string) => void;
+}
+
+const SearchBar: React.FC<SearchBarProps> = ({ onSearch }) => {
+  const [searchTerm, setSearchTerm] = useState<string>("");
+
+  const handleInputChange = (event: React.ChangeEvent<HTMLInputElement>) => {
+    const value = event.target.value;
+    setSearchTerm(value);
+    onSearch(value);
+  };
+
+  return (
+    <div className={styles.inputContainer}>
+      <BiSearch className={styles.iconSearch}></BiSearch>
+      <input
+        type="text"
+        placeholder="검색어를 입력하세요..."
+        value={searchTerm}
+        onChange={handleInputChange}
+        className={styles.searchBar}
+      />
+    </div>
+  );
+};
+
+export default SearchBar;
